+++ packages/ui-components/src/internals/CategoryBodyFacet/index.tsx
@@ -67,10 +67,9 @@ export const CategoryBodyFacet = compose(
           .toArray()
       );
     },
-  }),
-  sizeMe({ refreshRate: 50, refreshMode: 'debounce' })
-)(({ size, ...props }: any) => (
-  <div style={{ width: size.width }}>
+  })
+)((props: any) => (
+  <div>
     <Tree className={styles.wrap} {...props} isRoot selected has_children />
     {!props.isMobile &&
       props.childrenCount > props.config.maxItemsCount && (